Adds script to bulk download whole genomes
This may require a system wide installation of nextflow
( I don't think the conda env I provide has nextflow )

This script makes use of the ncbi_datasets command line tool
to bulk download a specified taxon/organism's data. To not
have a ridiculous download, I have the reference flag, but that
is toggleable.