Purpose of this workflow is simply to pull genomes
I have decided that I don't want this pipeline to do everything.
It will only pull down the genomes, and make a results file
that contains paths to the proper directory with all the genome files.

This will be used as the next input for another pipeline to perform tRNAscan.