Adds the beginning of a golden_gate toolkit
- codon_optimizer.py
To perform golden gate, I need to codon_optimize my sequence for the
organism im interested in. I was thinking that I would ask the user
for the donor genome as input. I would find the protein coding genes. Finally
I would get the average codon usage in the donor organism. I would do the same
for the recipient genome, then I would match up rare with rare, and
abundant with abundant codons.

- primer_maker
I was thinking I would use the programs online and replicate them, but
that might be a waste of time, because primers are just so easy to
make on 3rd party sites.

- seq_domesticator
I believe I actually got this somewhat working. It is supposed to
remove a given sequences Type IIS restriction sites seemlessly.